requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.3.7
chromadb>=0.4.18
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
pypdf2>=3.0.1
python-docx>=1.1.0
sentence-transformers>=2.2.2
numpy>=1.24.3
pandas>=2.0.3
sqlalchemy>=2.0.23
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
jinja2>=3.1.2
